Moved MessageTemplate, MessageTemplateAttachment, NetworkerReplyTrigger to static.py
