+++ mln/services/message.py
@@ -1,5 +1,5 @@
-from ..models.dynamic import Attachment, FriendshipStatus, Message, NetworkerReplyTrigger
-from ..models.static import MessageBody, MLNMessage
+from ..models.dynamic import Attachment, FriendshipStatus, Message
+from ..models.static import MessageBody, MLNMessage, NetworkerReplyTrigger
 from .friend import are_friends
 from .inventory import add_inv_item, remove_inv_item
